+++ CHANGELOG.md
@@ -9,11 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- race, gender, age, appearance and cutie mark fields for actors
+- added background fields for actors
+  - race
+  - gender
+  - age
+  - virtue ([#128](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/128))
+  - special talent
+  - appearance
+  - cutie mark
+  - personality
+  - social contacts
 - a limited character sheet for limited access to a character
   ([#11](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/11))
-- a virtue field for actors
-  ([#128](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/128))
 
 ### Changed